fs: Fixed a small bug in stream_write

+++ src/sys/keyboard.c
@@ -154,7 +154,7 @@ bool endkey = FALSE;
 bool shiftkey = FALSE;
 bool capslock = FALSE;
 uint8_t leds = 0;
-uint8_t key_pressed = 0;
+uint32_t key_pressed = 0;
 void kb_handle(uint8_t c)
 {
         key_pressed = 1;
